+++ tests/integration/bff-corss-project/bff-indep-client-app/api/lambda/context/index.ts
@@ -3,7 +3,7 @@ import { useHonoContext } from '@modern-js/server-runtime';
 export default async () => {
   const ctx = useHonoContext();
   const { res } = ctx;
-  res.append('x-id', '1');
+  res.headers.set('x-id', '1');
   return {
     message: 'Hello Modern.js',
   };