+++ packages/document/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @modern-js/main-doc
 
+## 3.1.5
+
+### Patch Changes
+
+- @modern-js/sandpack-react@3.1.5
+
+## 3.1.4
+
+### Patch Changes
+
+- @modern-js/sandpack-react@3.1.4
+
 ## 3.1.3
 
 ### Patch Changes